Modularisation
- renamed fast-decimal to fast-decimal-parent and set it to pom packaging
- created maven module fast-decimal-core
- moved all code and resources to fast-decimal-core
- cleaned up fast-decimal-parent
