+++ .gitignore
@@ -4,10 +4,7 @@ target/
 !**/src/test/**/target/
 
 ### IntelliJ IDEA ###
-.idea/modules.xml
-.idea/jarRepositories.xml
-.idea/compiler.xml
-.idea/libraries/
+.idea/
 *.iws
 *.iml
 *.ipr
@@ -36,4 +33,3 @@ build/
 
 ### Mac OS ###
 .DS_Store
-/.idea/vcs.xml
