+++ fast-decimal-core/pom.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>org.tlr</groupId>
+        <artifactId>fast-decimal-parent</artifactId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>fast-decimal-core</artifactId>
+    <name>Fast Decimal Core</name>
+    <description>Core implementation of the FastDecimal library</description>
+
+    <dependencies>
+        <!-- Test dependencies are inherited from parent -->
+    </dependencies>
+
+    <build>
+        <plugins>
+            <!-- Plugins are inherited from parent -->
+        </plugins>
+    </build>
+</project>
